cut fees by half
Fees increase the hashrate, in Bitcoin Static they increase the reward as well. Cutting fees reward in half will keep the supply unchanged, the miners will receive the exact same amount of money, as the burned half of the fees will be returned by the hashrate increase. It fixes an inflation bug.

Signed-off-by: Karol Trzeszczkowski <karol.trzeszczkowski@gmail.com>