Scale the default relay, mining, and wallet fee down by 100x
As discussed in chat with Licho:

Since fixtoshi's seem like they will hold more value -- and since the
circulating supply is so small now, we should probably make the default
fee be 0.01 fix/B (or 10 fix/kB).  This is basically 1 fix per 100B.

This commit also scales the dust theshold down by ~100x (since dust is
calculated off of minimum fee in a roundabout way).

This allows us to spend small amounts and also not pay high fees.

Additionally, the mining minimum fee was also set to the same as the
relay fee.