Set the default ports on all nets, set default RPC port as well
Made the default RPC port be 2136 (which is similar to how bitcoin's is
8332, one less than the p2p port of 8333). This is to avoid clashing with
an already-running bitcoind for BCH.

Also made the default ports for possible future testnet/scalenet/etc
be: p2p port xx2137 and RPC port xx2136, where xx is one of: 1,2,3,1
for test3, test4, scale.

Left the regrest ports unchanged (to prevent possible issues with the
python tests no longer working).