[qa] Correct the checkpoint for first post-UAHF block for testnet3
This is a follow-up to MR !1539 which corrected the UAHF checkpoint
block for mainnet.  This does similar for testnet3.  The old checkpoints
don't matter much now since we have since added more checkpoints after
them, but it's nice to be correct and consistent.  This is a purely
aesthetic/QA MR.

Test plan:

- ninja check-all
- Use a block explorer or BTC & BCH testnet3 node RPCs to verify that
  the existing hash for height 1155875 is
  `00000000f17c850672894b9a75b63a1e72830bbd5f4c8889b5c1a80e7faef138` on
   both chains (this is wrong as a checkpoint!).
- Use a block explorer or BTC & BCH testnet3 node RPCs to verify that
  the new hash in this MR for height 1155876 is
  `00000000000e38fef93ed9582a7df43815d5c2ba9fd37ef70c9a0ea4a285b8f5` and
  that this block exists only on BCH and not on BTC.