Fix inappropriate C-style usage of class uint256 in base58.cpp
Summary
---

This is a code quality commit -- it is inappropriate to use class uint256
as if it were a C-style array of uint8_t (even though the underlying
representation is indeed as such).  While the code in question is not
technically wrong as per the C++ standard, it is unsafe in case we ever
decide to change the representation of class uint256.  It is also
inelegant.

It is best to use the accessor methods .data() or .begin() to access
the underlying uint8_t bytes. This commit does just that.

Test Plan
---

- ninja all check-all