Upgrade Qt version to 5.15.3 for depends & gitian
This takes a bunch of changes from Core, heavily modified to work ok with
cmake, and upgrades the statically-linked Qt library we use for release
and/or depends from Qt 5.9.7 to Qt 5.15.3 (latest patched Qt5).
This brings lots of bugfixes and better support for HighDPI displays,
as well as correct OSX dark mode support (yay!).  By now, Qt 5.9.7 is
ancient and we really should be shipping the latest patched Qt with our
static binary releases.

Notes:

- We had to upgrade OpenSSL from 1.0.1 to 1.1.1n because Qt 5.15.3
  requires this library and refuses to work with older OpenSSL
  Ideally the openssl update would be a separate MR but the newer
  openssl refuses to work with our older Qt 5.9.8 !! Argh! So I had to
  roll these into the same MR.
- bison now is required on the host system when building using depends.
- OSX version works but requires clang-8.0.0 and OSX SDK 10.15
  (see: !1524).

Test Plan
---------

- build all platform with gitian
- verify hashes are reproducible
- test binaries on platforms that you can test them on